% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check.ps.and.zs.R
\name{check.ps.and.zs}
\alias{check.ps.and.zs}
\title{check.ps.and.zs}
\usage{
check.ps.and.zs(null.p.values, nonnull.p.values, plotQ = FALSE)
}
\arguments{
\item{null.p.values}{The p-values from the null (non-match) scores.}

\item{nonnull.p.values}{The p-values from the non-null (match) scores}

\item{plotQ}{Diagnostic plots?}
}
\value{
A list with indices of problem p/z-values and a list containing statistical test results for uniformity/normality.
}
\description{
Run diagnostic tests on p-values and z-values for both null (non-match) and non-null (match) scores.
}
\details{
Make a nice plot of overlapped histograms for each set of scores. NOTE: the y-axis is density not frequency.
}
\examples{
XXXX
}

