\name{empiricalPvalues}
\alias{empiricalPvalues}
\docType{package}
\title{
Simple function to comput p-values emiprically with respect to a set of null scores
}
\description{
Simple function to comput p-values emiprically with respect to a set of null scores
}
\usage{
empiricalPvalues(nullvec, vec)
}
\examples{
\dontrun{
empiricalPvalues(nullvec, vec)
}
}
