% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{hist.mode}
\alias{hist.mode}
\title{hist.mode}
\usage{
hist.mode(scores, bre, plotQ = FALSE)
}
\arguments{
\item{x}{data}
}
\description{
Handy mode function.
}
\details{
A handy mode function for a vector of data. Useful for poking around a troublesome non-null distribution 
of scores
}
\examples{
XXXX
}

