% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/null.gev.fit.R
\name{null.gev.fit}
\alias{null.gev.fit}
\title{null.gevd.fit}
\usage{
null.gev.fit(score.null.vec, standardizeQ = FALSE, plotQ = FALSE)
}
\arguments{
\item{score.null.vec}{Vector of null (non-match) similarity scores or transformed similarity scores}

\item{standardizeQ}{Whether or not to standardize the null scores}

\item{plotQ}{Diagnostic plots?}
}
\value{
list with the fitted parameters, fit info and chi-square goodness of fit test results
}
\description{
Generalized extreme value fit for null similarity scores.
}
\details{
The generalized extreme value distribution is a flexible three parameter distribution which often gives
a good fit to the null log similarity scores when these scores are on a scale of 0 to 1. Other than
standardization, this function assumes that if the user wants the scores to be transformed, they've
transformed them. This routine calls the \code{fgev} function from the evd package.
}
\examples{
XXXX
}
\references{
XXXX
}

