% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null.sn.fit.R
\name{null.sn.fit}
\alias{null.sn.fit}
\title{null.sn.fit}
\usage{
null.sn.fit(score.null.vec, standardizeQ = FALSE, plotQ = FALSE)
}
\arguments{
\item{score.null.vec}{Vector of null (non-match) similarity scores or transformed similarity scores}

\item{standardizeQ}{Whether or not to standardize the null scores}

\item{plotQ}{Diagnostic plots?}
}
\value{
list with the fitted parameters, fit info and chi-square goodness of fit test results
}
\description{
Skewed normal fit for null or transformed null similarity scores.
}
\details{
The skewed normal distribution is a flexible three parameter distribution which often gives 
a good fit to the null log similarity scores when these scores are on a scale of 0 to 1. Other than 
standardization, this function assumes that if the user wants the scores to be transformed, they've 
transformed them. This routine calls the \code{selm} function from the sn package.
}
\examples{
XXXX
}
\references{
XXXX
}

