% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.f.R
\name{process.f}
\alias{process.f}
\title{process.f}
\usage{
process.f(zscores, xmids, pct0, lam.mat)
}
\arguments{
\item{zscores}{Efron's z-scores derived from null and non-null score distributions}

\item{xmids}{Midpoints of z-score histogram cells}

\item{pct0}{Taken from locfdr package. Proportion of the z-score distribution used in fitting the null density f0(z) by central matching. If a 2-vector, e.g. pct0=c(0.25,0.60), the range [pct0[1], pct0[2]] is used. If a scalar, [pct0, 1-pct0] is used.}

\item{lam.mat}{Point estimate(s) of f(x) (lambda(x) in the Poission regression). Each column of lam.mat is an extimate of f(x). In a Bayesian regression context, each column is a draw of f(x) from the joint posterior distribution.}
}
\value{
A list consisting of:

delta0, null mean estimates

sig0, null standard dev estimates

p0, null prior estimates

f(x) estimates.
}
\description{
Computes estimates of the posterior null probabilities using f(x), the denominator of Bayes'
Theorem in Efron's two-groups methodology.
}
\details{
#Core function of the package. This algorithm take points making up the denominator
of Bayes theorem in in Efron's methodology f(x), computes
the likelihood function f0(x) = p(x|null) and (empirical)
estimate of the prior p0. One or many estimates of f(x) can be contained (as columns) in lam.mat.

The algorithim is largely taken from locfdr code. Follows basically
Efron 2006 Size, Power and False Discovery Rates

NOTE: Rows of lam.mat are f's generated by a Bayesian Poisson
regression onto the hisogram of counts generated by z-scores.
}
\examples{
XXXX
}
\references{
locfdr package by Efron, Turnbull and Narasimhan. 
Archived at http://cran.r-project.org/src/contrib/Archive/locfdr/
}

