% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.prep.R
\name{sampler.prep}
\alias{sampler.prep}
\title{sampler.prep}
\usage{
sampler.prep(p.values, num.bins = 120, degree = 7, interceptQ = FALSE,
  overdispersionQ = FALSE, sampler = c("jags", "stan"))
}
\arguments{
\item{XXXX}{}
}
\value{
XXXX
}
\description{
XXXX
}
\details{
XXXX
}
\examples{
XXXX
}
\references{
XXXX
}

