% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{scores.histograms}
\alias{scores.histograms}
\title{scores.histograms}
\usage{
scores.histograms(null.vec, nonnull.vec, xlim.min = NULL, xlim.max = NULL,
  ylim.max = NULL, xlab = NULL, main = NULL)
}
\arguments{
\item{null.vec}{Non-matching scores}

\item{nonnull.vec}{Matching scores}

\item{ylim.max}{Max on the y-axis. Helps beautify the overlayed histograms of null/non-null scores}

\item{xlab}{Optional x-axis label}

\item{main}{Optional plot title}
}
\description{
Make a nice plot of overlapped histograms for each set of scores.
}
\details{
Make a nice plot of overlapped histograms for each set of scores. NOTE: the y-axis is density not frequency.
}
\examples{
XXXX
}

