% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/smear.pz.R
\name{smear.extreme.nonnull.pvalues}
\alias{smear.extreme.nonnull.pvalues}
\title{smear.extreme.nonnull.pvalues}
\usage{
smear.extreme.nonnull.pvalues(non.null.pvalues, upper.set.pvalue = 1e-08,
  lower.set.pvalue = 3.5e-16, printQ = FALSE, plotQ = FALSE)
}
\arguments{
\item{XXXX}{}
}
\value{
XXXX
}
\description{
XXXX
}
\details{
deal with infinite z-values
}
\examples{
XXXX
}

