% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/smear.pz.R
\name{smear.extreme.nonnull.zvalues}
\alias{smear.extreme.nonnull.zvalues}
\title{smear.extreme.nonnull.zvalues}
\usage{
smear.extreme.nonnull.zvalues(non.null.pvalues, upper.set.zvalue = (-12),
  mu.factor = (-0.5), p.factor = 0.95, printQ = FALSE, plotQ = FALSE)
}
\arguments{
\item{XXXX}{}
}
\value{
XXXX
}
\description{
XXXX
}
\details{
deal with infinite z-values
}
\examples{
XXXX
}

