% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoomed.post.prob.plot.R
\name{zoomed.post.prob.plot}
\alias{zoomed.post.prob.plot}
\title{zoomed.post.prob.plot}
\usage{
zoomed.post.prob.plot(zscores, zbounds = c(NA, NA), prbounds = c(NA, NA),
  point.est.func, upper.est.func, lower.est.func, prob.scale = "percent",
  xlab = NULL, ylab = NULL, main = NULL)
}
\arguments{
\item{XXXX}{}
}
\value{
XXXX
}
\description{
XXXX
}
\details{
XXXX
}
\examples{
XXXX
}

