% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check.ps.and.zs.R
\name{check.ps.and.zs}
\alias{check.ps.and.zs}
\title{check.ps.and.zs}
\usage{
check.ps.and.zs(null.p.values = NULL, nonnull.p.values = NULL,
  printQ = FALSE, plotQ = FALSE)
}
\arguments{
\item{null.p.values}{Optional. The p-values from the null (non-match, KM) scores.}

\item{nonnull.p.values}{Optional. The p-values from the non-null (match, KNM) scores.}

\item{plotQ}{Diagnostic plots?}
}
\value{
A list with indices of problem p/z-values and a list containing statistical test results for
uniformity(p-values)/normality(z-values).
}
\description{
Run diagnostic tests on p-values and z-values for both null (non-match) and non-null (match) scores.
}
\details{
Check to see if: a. Null z-values are = +/- inf, Non-Null z-values are = -/+. If things are operating
according to plan (cf. Efron 2006) the Null z-values should be between +/- 3 with ~ 99% probability
(i.e. z_Null ~ N(0,1)  approximately). Extreme Null z-values are pathological. Check the KNM scores and/or
density fit to the KNM scores for weird-ness. It may be due to overly thin tails.
}
\examples{
XXXX
}

