% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make.fdr.functions.R
\name{make.fdr.functions}
\alias{make.fdr.functions}
\title{make.fdr.functions}
\usage{
make.fdr.functions(z.values, bin.midpoints, pct0 = 0.25, posterior.f.values,
  credibility.level = 0.95, interval.type = "equal.tail", plotQ = FALSE)
}
\arguments{
\item{XXXX}{}
}
\value{
XXXX
}
\description{
XXXX
}
\details{
XXXX
}
\examples{
XXXX
}

