% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{scores.histograms}
\alias{scores.histograms}
\title{scores.histograms}
\usage{
scores.histograms(null.vec, nonnull.vec, ylim.max)
}
\arguments{
\item{null.vec}{Non-matching scores}

\item{nonnull.vec}{Matching scores}

\item{ylim.max}{Max on the y-axis. Helps beautify the overlayed histograms of null/non-null scores}
}
\description{
Make a nice plot of overlapped histograms for each set of scores.
}
\details{
Make a nice plot of overlapped histograms for each set of scores.
}
\examples{
XXXX
}

